Fixed a refactoring bug from earlier, wasn't caught due to local caching.
